Handle PostgreSQL exceptions in the backend
The integrity constraint violation and the exception raised by a
constraint trigger (function) are probably due to the user data (e.g.
trying to modify a published document or leaving a NOT NULL field
unfilled), so these are seen as bad requests

A data error is probably a data validation error (e.g. filling a
string field with an int), but it's not clear if it's a client or a
server issue, so let's keep these as an internal server error for now